+++ src/pat/navigation/navigation.js
@@ -1,39 +1,33 @@
 define([
     "jquery",
-    "pat-registry",
+    "pat-base",
     "pat-parser",
     "pat-logger",
-], function($, registry, Parser, logger) {
+], function($, Base, Parser, logger) {
     var log = logger.getLogger("navigation");
     var parser = new Parser("navigation");
 
+    parser.addArgument("item-wrapper", "li");
     parser.addArgument("in-path-class", "navigation-in-path");
     parser.addArgument("current-class", "current");
 
-    var _ = {
+    return Base.extend({
         name: "navigation",
         trigger: "nav, .navigation, .pat-navigation",
-        current: null,
-        in_path: null,
         init: function($el, opts) {
             this.options = parser.parse($el, opts);
-            this.current = this.options.currentClass;
-            this.in_path = this.options.inPathClass;
-
-            var curpath = window.location.pathname;
-            log.debug("current path:", curpath);
-
+            var current = this.options.currentClass;
             // check whether to load
             if ($el.hasClass("navigation-load-current")) {
-                $el.find("a." + this.current, "."  + this.current + " a").click();
+                $el.find("a." + current, "."  + current + " a").click();
                 // check for current elements injected here
                 $el.on("patterns-injected-scanned", function(ev) {
                     var $target = $(ev.target);
-                    if ($target.is("a." + this.current))
+                    if ($target.is("a." + current))
                         $target.click();
-                    if ($target.is("." + this.current))
+                    if ($target.is("." + current))
                         $target.find("a").click();
-                    _._updatenavpath($el);
+                    this._updatenavpath($el);
                 }.bind(this));
             }
 
@@ -41,38 +35,63 @@ define([
             $el.on("patterns-inject-triggered", "a", function(ev) {
                 var $target = $(ev.target);
                 // remove all set current classes
-                $el.find("." + this.current).removeClass(this.current);
+                $el.find("." + current).removeClass(current);
                 // set current class on target
-                $target.addClass(this.current);
+                $target.addClass(current);
                 // If target's parent is an LI, also set current class there
-                $target.parent("li").addClass(this.current);
-                _._updatenavpath($el);
+                $target.parents(this.options.itemWrapper).first().addClass(current);
+                this._updatenavpath($el);
             }.bind(this));
 
+            var observer = new MutationObserver(this._initialSet.bind(this));
+            observer.observe($el[0], {
+                childList: true,
+                subtree: true,
+                attributes: false,
+                characterData: false,
+            });
+
+            this._initialSet();
+        },
+        _initialSet: function () {
+            var $el = this.$el;
+            var current = this.options.currentClass;
             // Set current class if it is not set
-            if ($el.find(this.current).length === 0) {
-                document.querySelectorAll("li a").forEach(function (it) {
+            if ($el[0].querySelectorAll('.' + current).length === 0) {
+                $el[0].querySelectorAll("a").forEach(function (it) {
                     var $a = $(it),
-                        $li = $a.parents("li:first"),
+                        $li = $a.parents(this.options.itemWrapper).first(),
                         url = $a.attr("href"),
                         path;
                     if (typeof url === "undefined") {
                         return;
                     }
-                    path = _._pathfromurl(url);
+                    path = this._pathfromurl(url);
                     log.debug("checking url:", url, "extracted path:", path);
-                    if (_._match(curpath, path)) {
+                    if (this._match(window.location.pathname, path)) {
                         log.debug("found match", $li);
-                        $li.addClass(this.current);
+                        $a.addClass(current);
+                        $li.addClass(current);
                     }
                 }.bind(this));
             }
-            _._updatenavpath($el);
+
+            // Set current class on item-wrapper, if not set.
+            if (
+                this.options.itemWrapper
+                && $el[0].querySelectorAll('.' + current).length > 0
+                && $el[0].querySelectorAll(this.options.itemWrapper + '.' + current).length === 0
+            ) {
+                $('.' + current, $el).parents(this.options.itemWrapper).first().addClass(current);
+            }
+
+            this._updatenavpath($el);
         },
         _updatenavpath: function($el) {
-            if (this.in_path) { return; }
-            $el.find(this.in_path).removeClass(this.in_path);
-            $el.find("li:not(." + this.current + "):has(." + this.current + ")").addClass(this.in_path);
+            var in_path = this.options.inPathClass;
+            if (! in_path) { return; }
+            $el.find('.' + in_path).removeClass(in_path);
+            $el.find(this.options.itemWrapper + ":not(." + this.options.currentClass + "):has(." + this.options.currentClass + ")").addClass(in_path);
         },
         _match: function(curpath, path) {
             if (!path) {
@@ -96,10 +115,5 @@ define([
             if (path.length === 1) return path[0];
             return path[1].split("/").slice(1).join("/");
         }
-    };
-    registry.register(_);
-    return _;
+    });
 });
-
-// jshint indent: 4, browser: true, jquery: true, quotmark: double
-// vim: sw=4 expandtab